refactor the security part and add them into utils and enhance the code logic
