+++ JS/sign-in/password-toggle.js
@@ -1,0 +1,12 @@
+function togglePassword() {
+  const password = document.getElementById('password');
+  const toggle = document.querySelector('.toggle-password');
+
+  if (password.type === "password") {
+    password.type = "text";
+    toggle.value = "🙈";
+  } else {
+    password.type = "password";
+    toggle.value = "👁️";
+  }
+}
