+++ sign-in.php
@@ -1,6 +1,10 @@
 <?php
 session_start();
+//! kicking out if authed
+require 'utils/auth-functions/auth-kick-out-of-log.php';
+
 require 'config/connection.php';
+
 if (isset($_POST['submit'])) {
 
   $email = mysqli_escape_string($con, $_POST['email']);
@@ -58,100 +62,85 @@ if (isset($_POST['submit'])) {
 <?php require('./template/header.php') ?>
 
 
-<?php if (empty($_SESSION['userName']) && empty($_SESSION['cUserEmail'])) { ?>
-  <div class="flex items-center justify-center min-h-screen bg-gradient-to-br from-red-500 via-orange-400 to-yellow-300">
-    <div class="bg-white bg-opacity-95 backdrop-blur-md p-8 rounded-lg shadow-lg max-w-sm w-full text-center">
-      <img src="./assets/Logo.png" alt="BurgerRiza Logo" class="w-20 mx-auto mb-4 rounded-full bg-white p-1 shadow-md">
-      <h2 class="text-2xl font-bold text-red-500 mb-2">Welcome Back!</h2>
-      <p class="text-gray-600 mb-6">Please log in to your account</p>
+<div class="flex items-center justify-center min-h-screen bg-gradient-to-br from-red-500 via-orange-400 to-yellow-300">
+  <div class="bg-white bg-opacity-95 backdrop-blur-md p-8 rounded-lg shadow-lg max-w-sm w-full text-center">
+    <img src="./assets/Logo.png" alt="BurgerRiza Logo" class="w-20 mx-auto mb-4 rounded-full bg-white p-1 shadow-md">
+    <h2 class="text-2xl font-bold text-red-500 mb-2">Welcome Back!</h2>
+    <p class="text-gray-600 mb-6">Please log in to your account</p>
 
-      <form action="<?php echo $_SERVER['PHP_SELF'] ?>" method="POST">
-        <div class="text-left">
-          <!-- User Inputs -->
-          <div class="flex justify-between">
-            <label for="Restaurant">
-              <input
-                type="radio"
-                id="Restaurant"
-                name="type"
-                value="R"
-                required
-                checked>
-              restaurant
-            </label>
-            <label for="Customers">
-              <input
-                type="radio"
-                id="Customers"
-                name="type"
-                value="C">
-              customer</label>
-          </div>
-
-          <section class="flex flex-col gap-y-4 py-4">
-            <label for="email" class="block text-sm font-medium text-gray-700">Email
-              <input
-                type="email"
-                id="email"
-                name="email"
-                value="<?php echo $_COOKIE['remember_email'] ?? '' ?>"
-                placeholder="Enter your email"
-                required
-                class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-red-500">
-            </label>
-
-            <div class="text-left">
-              <label for="password" class="block text-sm font-medium text-gray-700">Password
-                <div class="relative">
-                  <input
-                    type="password"
-                    id="password"
-                    name="password"
-                    placeholder="•••••••"
-                    required
-                    class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-red-500">
-                  <input
-                    class="absolute inset-y-0 right-3 flex items-center cursor-pointer text-gray-500 toggle-password"
-                    onclick="togglePassword()"
-                    type="button"
-                    value="👁️" />
-                </div>
-              </label>
-            </div>
-          </section>
+    <form action="<?php echo $_SERVER['PHP_SELF'] ?>" method="POST">
+      <div class="text-left">
+        <!-- User Inputs -->
+        <div class="flex justify-between">
+          <label for="Restaurant">
+            <input
+              type="radio"
+              id="Restaurant"
+              name="type"
+              value="R"
+              required
+              checked>
+            restaurant
+          </label>
+          <label for="Customers">
+            <input
+              type="radio"
+              id="Customers"
+              name="type"
+              value="C">
+            customer</label>
         </div>
 
-        <input
-          type="submit"
-          name="submit"
-          value="Log in"
-          class="w-full bg-red-500 text-white py-2 rounded-lg font-bold hover:bg-orange-400 transition-transform transform hover:scale-105" />
+        <section class="flex flex-col gap-y-4 py-4">
+          <label for="email" class="block text-sm font-medium text-gray-700">Email
+            <input
+              type="email"
+              id="email"
+              name="email"
+              value="<?php echo $_COOKIE['remember_email'] ?? '' ?>"
+              placeholder="Enter your email"
+              required
+              class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-red-500">
+          </label>
 
-        <label for="remember">
-          Remember me
-          <input type="checkbox" name="remember" <?php echo isset($_COOKIE['remember']) ? 'checked' : ''; ?> />
-        </label>
-      </form>
-      <?php echo $error ?? "" ?>
-      <p class="text-sm text-gray-600 mt-4">Don't have an account? <a href="sign-up.php" class="text-red-500 hover:underline">Sign Up</a></p>
-    </div>
+          <div class="text-left">
+            <label for="password" class="block text-sm font-medium text-gray-700">Password
+              <div class="relative">
+                <input
+                  type="password"
+                  id="password"
+                  name="password"
+                  placeholder="•••••••"
+                  required
+                  class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-red-500">
+                <input
+                  class="absolute inset-y-0 right-3 flex items-center cursor-pointer text-gray-500 toggle-password"
+                  onclick="togglePassword()"
+                  type="button"
+                  value="👁️" />
+              </div>
+            </label>
+          </div>
+        </section>
+      </div>
+
+      <input
+        type="submit"
+        name="submit"
+        value="Log in"
+        class="w-full bg-red-500 text-white py-2 rounded-lg font-bold hover:bg-orange-400 transition-transform transform hover:scale-105" />
+
+      <label for="remember">
+        Remember me
+        <input type="checkbox" name="remember" <?php echo isset($_COOKIE['remember']) ? 'checked' : ''; ?> />
+      </label>
+    </form>
+    <?php echo $error ?? "" ?>
+    <p class="text-sm text-gray-600 mt-4">Don't have an account? <a href="sign-up.php" class="text-red-500 hover:underline">Sign Up</a></p>
   </div>
-  <script>
-    function togglePassword() {
-      const password = document.getElementById('password');
-      const toggle = document.querySelector('.toggle-password');
+</div>
 
-      if (password.type === "password") {
-        password.type = "text";
-        toggle.value = "🙈";
-      } else {
-        password.type = "password";
-        toggle.value = "👁️";
-      }
-    }
-  </script>
-<?php } else {
-  header('Location:index.php');
-} ?>
+<!-- script of function  -->
+<script src="JS/sign-in/password-toggle.js"></script>
 
 <?php require('./template/footer.php') ?>
