+++ utils/auth-functions/auth-kick-out-of-log.php
@@ -1,0 +1,4 @@
+<?php if (!empty($_SESSION['userName'])) {
+  header('Location:index.php');
+  die;
+}
