+++ utils/auth-functions/customer-page/kick-res-owner.php
@@ -1,0 +1,5 @@
+<?php
+if ($_SESSION['Role'] !== "C") {
+  header('Location:index.php');
+  die;
+}
